requirements.txt
absl-py==0.7.0
aniso8601==8.0.0
astor==0.7.1
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi
chardet==4.0.0
click==7.1.2
colorama==0.4.1
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.10
ImageHash==4.1.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
opencv-contrib-python==3.3.0.10
opencv-python==3.3.0.10
pandas==1.1.0
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.4.1
pipreqs==0.4.10
progressbar2==3.51.4
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
pyclustering==0.9.3.1
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-utils==2.4.0
pytz
PyWavelets==1.1.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.25.1
scikit-learn==0.23.1
scipy==1.2.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sympy==1.4
tensorboard==1.11.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
threadpoolctl==2.1.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
wincertstore==0.2
yarg==0.1.9
